after modified mvc for attraction and user
